Add PdfViewerService
